requirements.txt
GitPython
pathspec
